move quantum_data to new data dir
